New data strategy: try cache, /tmp, cms
Load data from cache (memory).
If data is not yet in cache, load it from temp directory on file system (os.tmpdir()) and put it in cache.
If data is not yet on file system either, load it from the cms (dump), save it to the file system, and put it in cache.